+++ api_gateway/client/src/app/models/execution/consoleLog.ts
@@ -1,22 +1,22 @@
 export class ConsoleLog {
 	message: string;
 
-	action_name: string;
-
-	app_name: string;
-
-	workflow: string;
-
-	level: string;
+	// action_name: string;
+  //
+	// app_name: string;
+  //
+	// workflow: string;
+  //
+	// level: string;
 
 	toNewConsoleLog(): ConsoleLog {
 		const out = new ConsoleLog();
 
 		out.message = this.message;
-		out.action_name = this.action_name;
-		out.app_name = this.app_name;
-		out.workflow = this.workflow;
-		out.level = this.level;
+		// out.action_name = this.action_name;
+		// out.app_name = this.app_name;
+		// out.workflow = this.workflow;
+		// out.level = this.level;
 
 		return out;
 	}