+++ app_sdk/walkoff_app_sdk/app_base.py
@@ -38,7 +38,7 @@ class HTTPStream:
     async def write(self, message):
         data = {"message": message}
         params = {"workflow_execution_id": self.execution_id}
-        url = f"{API_GATEWAY_URI}/api/streams/console/log"
+        url = f"{API_GATEWAY_URI}/api/streams/console/logger"
 
         await self.session.post(url, json=data, params=params)
 
@@ -138,6 +138,7 @@ class AppBase:
                             else:
                                 params[p.name] = p.value
                         result = await func(**params)
+
                     action_result = NodeStatusMessage.success_from_node(action, action.execution_id, result=result)
                     self.logger.debug(f"Executed {action.label}-{action.id_} with result: {result}")
 
@@ -147,8 +148,8 @@ class AppBase:
                                                                         result="Action not callable")
 
             except Exception as e:
-                action_result = NodeStatusMessage.failure_from_node(action, action.execution_id, result=repr(e))
                 self.logger.exception(f"Failed to execute {action.label}-{action.id_}")
+                action_result = NodeStatusMessage.failure_from_node(action, action.execution_id, result=repr(e))
 
         else:
             self.logger.error(f"App {self.__class__.__name__} has no method {action.name}")