+++ common/message_types.py
@@ -21,6 +21,7 @@ def message_load(obj):
 
 class MessageJSONDecoder(json.JSONDecoder):
     """ A custom decoder for decoding JSON strings to Message types. """
+
     def __init__(self, *args, **kwargs):
         json.JSONDecoder.__init__(self, object_hook=self.object_hook, *args, **kwargs)
 
@@ -42,12 +43,21 @@ class MessageJSONDecoder(json.JSONDecoder):
 
 class MessageJSONEncoder(json.JSONEncoder):
     """ A custom encoder for encoding Message types to JSON strings. """
+
     def default(self, o):
         if isinstance(o, NodeStatusMessage):
-            return {"name": o.name, "node_id": o.node_id, "label": o.label, "app_name": o.app_name,
-                    "execution_id": o.execution_id, "result": o.result, "status": o.status,
-                    "started_at": o.started_at, "completed_at": o.completed_at, "combined_id": o.combined_id,
-                    "arguments": o.arguments}
+            r = {"name": o.name, "node_id": o.node_id, "label": o.label, "app_name": o.app_name,
+                 "execution_id": o.execution_id, "result": o.result, "status": o.status,
+                 "started_at": o.started_at, "completed_at": o.completed_at, "combined_id": o.combined_id,
+                 "parameters": o.parameters}
+
+            try:
+                json.dumps(o.result)
+            except (TypeError, ValueError):
+                r["result"] = f"Node returned result of type '{type(o.result)}' which is not JSON serializable."
+                r["status"] = StatusEnum.FAILURE
+            finally:
+                return r
 
         elif isinstance(o, WorkflowStatusMessage):
             return {"execution_id": o.execution_id, "workflow_id": o.workflow_id, "name": o.name, "status": o.status,
@@ -144,10 +154,10 @@ class WorkflowStatusMessage(object):
 
 class NodeStatusMessage(object):
     """ Class that formats a NodeStatusMessage message. """
-    __slots__ = ("name", "node_id", "label", "app_name", "execution_id", "arguments", "combined_id", "result",
+    __slots__ = ("name", "node_id", "label", "app_name", "execution_id", "parameters", "combined_id", "result",
                  "status", "started_at", "completed_at")
 
-    def __init__(self, name, node_id, label, app_name, execution_id, combined_id=None, arguments=None, result=None,
+    def __init__(self, name, node_id, label, app_name, execution_id, combined_id=None, parameters=None, result=None,
                  status=None, started_at=None, completed_at=None):
         self.name = name
         self.node_id = node_id
@@ -156,41 +166,43 @@ class NodeStatusMessage(object):
         self.execution_id = execution_id
         self.combined_id = combined_id if combined_id is not None else ':'.join((node_id, execution_id))
         self.result = result
-        self.arguments = arguments
+        self.parameters = parameters
         self.status = status
         self.started_at = started_at
         self.completed_at = completed_at
 
     @classmethod
-    def from_node(cls, node, execution_id, result=None, status=None, started_at=None, completed_at=None):
+    def from_node(cls, node, execution_id, result=None, status=None, started_at=None, completed_at=None, parameters=None):
         return cls(node.name, node.id_, node.label, node.app_name, execution_id, result=result,
-                   status=status, started_at=started_at, completed_at=completed_at)
+                   status=status, started_at=started_at, completed_at=completed_at, parameters=parameters)
 
     @classmethod
-    def pending_from_node(cls, node, execution_id):
-        return NodeStatusMessage.from_node(node, execution_id, status=StatusEnum.PENDING)
+    def pending_from_node(cls, node, execution_id, parameters=None):
+        return NodeStatusMessage.from_node(node, execution_id, status=StatusEnum.PENDING, parameters=parameters)
 
     @classmethod
-    def executing_from_node(cls, node, execution_id):
+    def executing_from_node(cls, node, execution_id, parameters=None):
         started_at = datetime.datetime.now()
-        return NodeStatusMessage.from_node(node, execution_id, started_at=started_at, status=StatusEnum.EXECUTING)
+        return NodeStatusMessage.from_node(node, execution_id, started_at=started_at, status=StatusEnum.EXECUTING,
+                                           parameters=parameters)
 
     @classmethod
-    def success_from_node(cls, node, execution_id, result):
+    def success_from_node(cls, node, execution_id, result, parameters=None):
         completed_at = datetime.datetime.now()
         return NodeStatusMessage.from_node(node, execution_id, result=result, completed_at=completed_at,
-                                           status=StatusEnum.SUCCESS)
+                                           status=StatusEnum.SUCCESS, parameters=parameters)
 
     @classmethod
-    def failure_from_node(cls, node, execution_id, result):
+    def failure_from_node(cls, node, execution_id, result, parameters=None):
         completed_at = datetime.datetime.now()
         return NodeStatusMessage.from_node(node, execution_id, result=result, completed_at=completed_at,
-                                           status=StatusEnum.FAILURE)
+                                           status=StatusEnum.FAILURE, parameters=parameters)
+
     @classmethod
-    def aborted_from_node(cls, node, execution_id):
+    def aborted_from_node(cls, node, execution_id, parameters=None):
         completed_at = datetime.datetime.now()
         return NodeStatusMessage.from_node(node, execution_id, result=None, completed_at=completed_at,
-                                           status=StatusEnum.ABORTED)
+                                           status=StatusEnum.ABORTED, parameters=parameters)
 
 
 class TriggerMessage(object):