+++ key_creation.py
@@ -1,0 +1,3 @@
+import uuid
+
+print(str(uuid.uuid4().hex))